feat: Add Decide Function Zero
Add decide function 0 that takes an array of 2D coordinates and a double length. Returns true if two consecutive coordinates are at a distance of miminum length, otherwise false.

Also adds test that checks the function for various inputs.

Lastly adds a small change to .gitignore for Eclipse auto-generated files.