Issue #33 Add LIC function one (#82)
* feat/#33 Add LIC function one

Add LIC function one that takes an array of coordinates and a radius variable. It tests all three consecutive coordinates to see if they cannot fit within a circle of a given radius.

Also adds test functions for LIC function one.

* fix/#86 Removed unused and duplicate imports, trying to solve broken build

The reason for the build not passing on Travis seams to be double imports of the MathHelper class. 
Co-authored-by: Nils Streijffert <nils.streijffert@gmail.com>