feat/#34 Add LIC function three
Add LIC function three that takes an area and an array of coordinates. Checks if there are three consecutive coordinates that form a triangle with an area greater than the one given.

Also adds test functions for LIC functions three.

Add a help function in MathHelper that checks if two floating point numbers has a difference smaller than EPSILON. Used to check if two numbers can be considered equal in order to prevent rounding errors.