feat/fix/#98 Add LIC function fourteen
Add first version of LIC function fourteen  implementation. The functions takes two integers that specify the number of points separating three consecutive points respectively, two given areas, and an array of coordinates. If three points can be found that satisfy the consecutive point distance and a triangle is found with an area greater than area1, and one triangle with area smaller than area2, return true. Otherwise false.

Also adds one basic test for the function.

Fix a small bug in MathHelper regarding the equal function so that the order of the input variables doesn't matter.