fix/refactor/#98/#104 Fix triangle area calculation
Fix a bug in the calculation of the area of a triangle so that it recognizes when three points form a line and not a triangle.

Refactor the triangle area calculation by adding a function for it to MathHelper that is used by both LIC function three and fourteen.

Also adds a few more tests for LIC function fourteen.