Issue #98 and #104 Implement LIC function fourteen and fix bug in triangle area calculation (#105)
* feat/fix/#98 Add LIC function fourteen

Add first version of LIC function fourteen  implementation. The functions takes two integers that specify the number of points separating three consecutive points respectively, two given areas, and an array of coordinates. If three points can be found that satisfy the consecutive point distance and a triangle is found with an area greater than area1, and one triangle with area smaller than area2, return true. Otherwise false.

Also adds one basic test for the function.

Fix a small bug in MathHelper regarding the equal function so that the order of the input variables doesn't matter.

* fix/refactor/#98/#104 Fix triangle area calculation fixes #104 

Fix a bug in the calculation of the area of a triangle so that it recognizes when three points form a line and not a triangle.

Refactor the triangle area calculation by adding a function for it to MathHelper that is used by both LIC function three and fourteen.

Also adds a few more tests for LIC function fourteen.

* fix/#98 Add further test for LIC function fourteen fixes #98 

Add a few more tests for LIC function fourteen.

* doc/#106 Add JavaDoc for LIC function three and fourteen fixes #106

Add JavaDoc comments to LIC function three and fourteen.